Move all ref selectors to block editor
