+++ docs/reference-guides/data/data-core-editor.md
@@ -435,10 +435,6 @@ _Related_
 
 -   getInserterItems in core/block-editor store.
 
-### getInserterSearchInputRef
-
-Undocumented declaration.
-
 ### getLastMultiSelectedBlockClientId
 
 _Related_
